requirements.txt
streamlit
torch
torchaudio
transformers
openai-whisper
speechrecognition
pyaudio
gtts 
sentence-transformers
faiss-cpu
numpy
gradio
openai
python-dotenv
requests